test(server): split the workspace diff lane
Workspace diff and commit wire tests were among the longest jobs in the monolithic server binary and competed with unrelated provider and runtime tests for Git processes. Move the public HTTP group into the existing workspace package and bound it to eight concurrent Git-heavy tests so the Go scheduler can run it beside the root package without overloading the host.

Validation: GOMAXPROCS=24 go test ./internal/server/workspacetest -parallel=8 -shuffle=on completed in 10.3s. The root package reached 179.2s under the same CPU/parallel settings; one unrelated private cache-invalidation test was load-sensitive in that run and passed in focused reproduction.
